backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.2",
    "gitpython>=3.1.46",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "langchain>=1.2.9",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.0",
]
